# www/docs/openapi.yaml
openapi: 3.0.0
info:
  description: >-
    This is the API specication for the [Wazigte Edge](https://github.com/Waziup/wazigate-edge).
  version: 1.0.0
  title: Wazigate Edge
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: johann.forster@waziup.org
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/apps':
    get:
      tags:
        - Apps
      summary: List Apps
      description: Apps are docker containers. They can be active (running) or inactive (stopped).
      parameters:
        - name: internal
          in: query
          description: Include internal apps.
          required: false
          schema:
            type: boolean
            example: false
        - name: updates
          in: query
          description: Check if a newer app version (a newer image from the Docker Hub) is available. This will query the docker registry.
          required: false
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: Successful Operation.
        '500':
          description: The was an internal error maybe because the docker daemon did not respond or because there was no internet connection to perform the updates check.
    post:
      tags:
        - Apps
      summary: Install new Apps
      description: Fetch the given image from the docker repository. Calling this api will use a `docker pull` internally and create a container from the image.
      requestBody:
        description: Image data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: Image name
                  example: "waziup/wazigate-lora:V1.2"
      responses:
        '200':
          description: Successful Operation.
        '400':
          description: There is no image with that name.
        '500':
          description: The was an internal error maybe because the docker daemon did not respond or because there was no internet connection to perform this action.
  '/apps/{name}':
    parameters:
      - name: name
        in: path
        description: Name of the app.
        required: true
        schema:
          type: string
          example: "wazigate-lora"
    get:
      tags:
        - Apps
      summary: Get App data
      description: Apps are docker containers.
      responses:
        '200':
          description: Successful Operation.
        '404':
          description: There is no app (Docker Container) with that name.
    post:
      tags:
        - Apps
      summary: Manage an App, Start & Stop
      description: Start or stop an App.
      parameters:
        - name: start
          in: query
          description: Start an app.
          required: false
          schema:
            type: boolean
            example: false
        - name: stop
          in: query
          description: Stop an app.
          required: false
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: Successful Operation.
        '404':
          description: There is no container with that name.
        '500':
          description: The was an internal error maybe because the docker daemon did not respond.
    delete:
      tags:
        - Apps
      summary: Delete an App
      description: Apps are docker containers. The App must be stopped to perform this operation. This action wil delete the Docker Container and the Docker Image.
      responses:
        '200':
          description: Successful Operation.
        '404':
          description: There is no image with that name.
        '500':
          description: There was an internal error maybe because the docker daemon did not respond or it refused to delete the container or the image.
  '/apps/{name}/{path}':
    parameters:
      - name: name
        in: path
        description: Name of the App.
        required: true
        schema:
          type: string
          example: "wazigate-lora"
      - name: path
        in: path
        description: Proxy path.
        required: true
        schema:
          type: string
          example: "index.html"
    get:
      tags:
        - Apps
      summary: Apps Proxy
      description:
        Get app files through the apps proxy.

        All apps should host a http server that is listening on `WAZIGATE_ADDR` which is a environment variable given to the container. Wazigate-compatible containers are required to expose a `package.json` file on http root level.
      responses:
        '200':
          description: Successful Operation.
        '404':
          description: There is no app with that name or the app returned a 404 error.
        '500':
          description: The was an internal error.
      
      
# System #######################################################################################################
...